feat: Restructure ETL
Add the ability to create a new SQLite database in the load_data function based
on the configuration file provided. This functionality is necessary to ensure
that data can be properly loaded into the destination database.

The new database_setup function checks the configuration file for database
creation instructions and proceeds to create the database if required. This
ensures that the destination database is properly set up before data loading
begins.